+++ CHANGELOG.md
@@ -13,11 +13,28 @@
 ### Fixed
 - No fixes
 
+## [0.0.2] - 2020-07-03
+
+### Added
+
+#### Customize For Operating System Modifier
+
+A modifier that allows adding additional modifiers or customizations to a view based on the current operating system. This allows us to avoid splitting up view builders to be able to use  `#if os(tvOS)`.
+
+```
+Text("Hello World!")
+   .font(Font.system(size: 12.0))
+   .customize(for: .tvOS, modifier: {
+       $0.font(Font.system(size: 20.0))
+   })
+```
+
 ## [0.0.1] - 2020-07-03
 
 ### Added
-- Theme
-    - A theme system based on `@Environment` that can be accessed from SwiftUI views. [@Marxon13](https://marxon13.com)
-    - System colors for iOS, macOS, tvOS, and watchOS to `Theme`. [@Marxon13](https://marxon13.com)
-    - Standardized semantic colors defined by the design system to `Theme`. [@Marxon13](https://marxon13.com)
+
+#### Theme
+- A theme system based on `@Environment` that can be accessed from SwiftUI views. [@Marxon13](https://marxon13.com)
+- System colors for iOS, macOS, tvOS, and watchOS to `Theme`. [@Marxon13](https://marxon13.com)
+- Standardized semantic colors defined by the design system to `Theme`. [@Marxon13](https://marxon13.com)